# openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/rag/stream/ask/news": {
      "post": {
        "tags": ["rag"],
        "summary": "串流回答新聞相關",
        "operationId": "ask_stream_news_api_v1_rag_stream_ask_news_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AskRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/news/": {
      "get": {
        "tags": ["news"],
        "summary": "取得新聞列表",
        "operationId": "read_news_api_v1_news__get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "title": "Start" }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "title": "End" }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "title": "Tag" }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "title": "Keyword" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/NewsCard" },
                  "title": "Response Read News Api V1 News  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/news/stats": {
      "get": {
        "tags": ["news"],
        "summary": "取得每月新聞統計資訊",
        "operationId": "read_news_stats_api_v1_news_stats_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "title": "Start" }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "title": "End" }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "title": "Tag" }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "title": "Keyword" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NewsStats" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/news/wordcloud": {
      "get": {
        "tags": ["news"],
        "summary": "取得新聞關鍵字雲",
        "operationId": "read_news_wordcloud_api_v1_news_wordcloud_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "title": "Start" }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "title": "End" }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "title": "Tag" }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "title": "Keyword" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WordCloud" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/news/artist-volume": {
      "get": {
        "tags": ["news"],
        "summary": "取得藝人新聞量",
        "operationId": "read_artist_volume_api_v1_news_artist_volume_get",
        "parameters": [
          {
            "name": "names",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": { "type": "string" },
              "title": "Names"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ArtistVolumePoint" },
                  "title": "Response Read Artist Volume Api V1 News Artist Volume Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtistVolumePoint": {
        "properties": {
          "artist_name": { "type": "string", "title": "Artist Name" },
          "data": {
            "additionalProperties": { "type": "integer" },
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": ["artist_name", "data"],
        "title": "ArtistVolumePoint"
      },
      "AskRequest": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question",
            "example": "可以推薦給我幾首浪漫的歌曲嗎，並說明推薦的原因"
          }
        },
        "type": "object",
        "required": ["question"],
        "title": "AskRequest"
      },
      "DailyCount": {
        "properties": {
          "date": { "type": "string", "title": "Date" },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["date", "count"],
        "title": "DailyCount"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NewsCard": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "title": { "type": "string", "title": "Title" },
          "date": { "type": "string", "title": "Date" },
          "tag": { "type": "string", "title": "Tag" },
          "image": { "type": "string", "title": "Image" },
          "link": { "type": "string", "title": "Link" },
          "content": { "type": "string", "title": "Content" }
        },
        "type": "object",
        "required": ["id", "title", "date", "tag", "image", "link", "content"],
        "title": "NewsCard"
      },
      "NewsStats": {
        "properties": {
          "counts": {
            "items": { "$ref": "#/components/schemas/DailyCount" },
            "type": "array",
            "title": "Counts"
          }
        },
        "type": "object",
        "required": ["counts"],
        "title": "NewsStats"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "WordCloud": {
        "properties": {
          "top_words": {
            "items": { "$ref": "#/components/schemas/WordStat" },
            "type": "array",
            "title": "Top Words"
          }
        },
        "type": "object",
        "required": ["top_words"],
        "title": "WordCloud"
      },
      "WordStat": {
        "properties": {
          "word": { "type": "string", "title": "Word" },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["word", "count"],
        "title": "WordStat"
      }
    }
  }
}